DecimalQuantiles: Move QuantileSerializationType to separate header
Previous one was autogenerated
