+++ src/include/duckdb/core_functions/aggregate/holistic_functions.hpp
@@ -84,12 +84,4 @@ struct ReservoirQuantileFun {
 	static AggregateFunctionSet GetFunctions();
 };
 
-enum class QuantileSerializationType : uint8_t {
-	NON_DECIMAL = 0,
-	DECIMAL_DISCRETE,
-	DECIMAL_DISCRETE_LIST,
-	DECIMAL_CONTINUOUS,
-	DECIMAL_CONTINUOUS_LIST
-};
-
 } // namespace duckdb
